Rename StatePresentOffer to ProductInfo
