+++ src/ui/components/product-info.svelte
