+++ src/ui/localization/supportedLanguages.ts
@@ -59,16 +59,16 @@ export enum LocalizationKeys {
   PeriodsMonthFrequencyPlural = "periods.monthFrequencyPlural",
   PeriodsYearFrequencyPlural = "periods.yearFrequencyPlural",
   PeriodsDayFrequencyPlural = "periods.dayFrequencyPlural",
-  StatePresentOfferProductTitle = "state_present_offer.product_title",
-  StatePresentOfferProductDescription = "state_present_offer.product_description",
-  StatePresentOfferProductPrice = "state_present_offer.product_price",
-  StatePresentOfferFreeTrialDuration = "state_present_offer.free_trial_duration",
-  StatePresentOfferPriceAfterFreeTrial = "state_present_offer.price_after_free_trial",
-  StatePresentOfferPriceTotalDueToday = "state_present_offer.total_due_today",
-  StatePresentOfferSubscribeTo = "state_present_offer.subscribe_to",
-  StatePresentOfferRenewalFrequency = "state_present_offer.renewal_frequency",
-  StatePresentOfferContinuesUntilCancelled = "state_present_offer.continues_until_cancelled",
-  StatePresentOfferCancelAnytime = "state_present_offer.cancel_anytime",
+  ProductInfoProductTitle = "state_present_offer.product_title",
+  ProductInfoProductDescription = "state_present_offer.product_description",
+  ProductInfoProductPrice = "state_present_offer.product_price",
+  ProductInfoFreeTrialDuration = "state_present_offer.free_trial_duration",
+  ProductInfoPriceAfterFreeTrial = "state_present_offer.price_after_free_trial",
+  ProductInfoPriceTotalDueToday = "state_present_offer.total_due_today",
+  ProductInfoSubscribeTo = "state_present_offer.subscribe_to",
+  ProductInfoRenewalFrequency = "state_present_offer.renewal_frequency",
+  ProductInfoContinuesUntilCancelled = "state_present_offer.continues_until_cancelled",
+  ProductInfoCancelAnytime = "state_present_offer.cancel_anytime",
   StateNeedsAuthInfoEmailStepTitle = "state_needs_auth_info.email_step_title",
   StateNeedsAuthInfoEmailInputLabel = "state_needs_auth_info.email_input_label",
   StateNeedsAuthInfoEmailInputPlaceholder = "state_needs_auth_info.email_input_placeholder",
